fix2

+++ be/src/vec/exec/format/parquet/parquet_common.h
@@ -48,12 +48,11 @@ struct RowRange {
 
 #pragma pack(1)
 struct ParquetInt96 {
-    uint64_t lo; // time of nanoseconds in a day
-    uint32_t hi; // days from julian epoch
+    int64_t lo; // time of nanoseconds in a day
+    int32_t hi; // days from julian epoch
 
     inline int64_t to_timestamp_micros() const {
-        return ((int32_t)hi - JULIAN_EPOCH_OFFSET_DAYS) * MICROS_IN_DAY +
-               int64_t(lo) / NANOS_PER_MICROSECOND;
+        return (hi - JULIAN_EPOCH_OFFSET_DAYS) * MICROS_IN_DAY + lo / NANOS_PER_MICROSECOND;
     }
     inline __int128 to_int128() const {
         __int128 ans = 0;
